python/pyproject.toml
[project]
name = "selfspy"
version = "1.0.0"
description = "A tool for monitoring and analyzing your computer activity"
authors = [{name = "nuin", email = "nuin@genedrift.org"}]
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
dependencies = [
    "SQLAlchemy>=2.0.25",
    "aiosqlite>=0.19.0",
    "pynput>=1.7.6",
    "cryptography>=41.0.7",
    "keyring>=24.3.0",
    "python-dateutil>=2.8.2",
    "typer>=0.9.0",
    "appdirs>=1.4.4",
    "rich>=13.7.0",
    "structlog>=24.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "plotly>=5.18.0",
    "pandas>=2.1.4",
    "pydantic-settings>=2.1.0",
    "greenlet>=3.0.1",
    "aiofiles>=23.2.0",
    "pycryptodome>=3.19.0",
]

[project.optional-dependencies]
macos = [
    "pyobjc-framework-Quartz",
    "pyobjc-framework-ApplicationServices",
]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.1.11",
    "pre-commit>=3.6.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.3",
    "black>=23.12.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "ruff>=0.1.11",
    "pre-commit>=3.6.0",
]

[project.scripts]
selfspy = "src.cli:app"
selfstats = "src.stats:app"
selfviz = "src.enhanced_stats:app"
selfterminal = "src.terminal_stats:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-ra -q --cov=src"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
check_untyped_defs = true